Version 2.0.0
1. Security updated.
2. From now on e-Wallet IS FULLY FUNCTIONAL.
**Now it is more secure. three layer protection for eWallet.

+++ e-Wallet_API/e_wallet/singleHistory.php
@@ -14,40 +14,34 @@
     if($link == null){
         http_response_code(404);
         echo json_encode(array("status" => "Connection problem on server"));
-    }else if($data == null || !property_exists($data, 'id') || !property_exists($data, 'pass') || ! property_exists($data, 'key')){
+    }else if($data == null || !property_exists($data, 'id') || !property_exists($data, 'key') || !property_exists($data, 'pass') || !property_exists($data, 'trid')){
         $conObg->detach();
         echo "Get Lost";
-    }else if( !$kit->test_input($data->id) || !$kit->test_input($data->pass) || !$kit->test_input($data->key) ){
+    }else if( !$kit->test_input($data->id) || !$kit->test_input($data->pass) || !$kit->test_input($data->key) || !$kit->test_input($data->trid)){
         $conObg->detach();     
         echo "You  fool, Get Lost";
     }
     else{ // do everything here, key is needed to intify request source.        
-        $qry = "SELECT name, address, hasPass, secQues as 'qs' FROM `user_data` WHERE nsuId = $data->id 
+        $qry = "SELECT hashPin FROM `wallet` WHERE nsuId = $data->id 
                                        and
               (SELECT count(*) FROM `req_data` WHERE appKey = $data->key) = 1";
        // this qry will check app key, if valid then check id
 
         $res = mysqli_fetch_all(mysqli_query($link, $qry), MYSQLI_ASSOC);
-        if($res != null){
-            if(password_verify($data->pass, $res[0]['hasPass'])){
-                $qry = "SELECT onOrOf FROM `wallet` WHERE nsuId = $data->id ";
-                $res1 = mysqli_fetch_all(mysqli_query($link, $qry), MYSQLI_ASSOC);
-                $conObg->detach();
-                $response = array(
-                    "status"=> "ok",
-                    "name" => $res[0]['name'],
-                    "add"  => $res[0]['address'],
-                    "qstn" => $res[0]['qs'],
-                    "flag" => ($res1 == null)? null : $res1[0]['onOrOf']
-                );
+        if($res != null && password_verify($data->pass, $res[0]['hashPin'])){
+                       
+            $qry = "SELECT list FROM `history` WHERE trid = '$data->trid'";
+            $res = mysqli_fetch_all(mysqli_query($link, $qry), MYSQLI_ASSOC);
+            $conObg->detach();
+            $data =  $res[0]['list'];
+            if($res != null){
                 http_response_code(200);
-                echo json_encode($response);
+                echo json_encode(array("status"=> "ok","data"=> "$data"));
             }else{
                 http_response_code(200);
-                echo json_encode(array('status'=> 'invalid'));
+                echo json_encode(array('status'=> 'not ok'));
             }
-
-            
+                        
         }else{  // invalid app key
             $conObg->detach();
              echo "Get Lost, you fool.";
