+++ e-Wallet_API/e_wallet/viewAllHis.php
@@ -14,38 +14,36 @@
     if($link == null){
         http_response_code(404);
         echo json_encode(array("status" => "Connection problem on server"));
-    }else if($data == null || !property_exists($data, 'id') || !property_exists($data, 'pass') || ! property_exists($data, 'key')){
+    }else if($data == null || !property_exists($data, 'id') || !property_exists($data, 'key') || !property_exists($data, 'pass')){
         $conObg->detach();
         echo "Get Lost";
-    }else if( !$kit->test_input($data->id) || !$kit->test_input($data->pass) || !$kit->test_input($data->key) ){
+    }else if( !$kit->test_input($data->id) || !$kit->test_input($data->pass) || !$kit->test_input($data->key)){
         $conObg->detach();     
         echo "You  fool, Get Lost";
     }
     else{ // do everything here, key is needed to intify request source.        
-        $qry = "SELECT name, address, hasPass, secQues as 'qs' FROM `user_data` WHERE nsuId = $data->id 
+        $qry = "SELECT hashPin FROM `wallet` WHERE nsuId = $data->id 
                                        and
               (SELECT count(*) FROM `req_data` WHERE appKey = $data->key) = 1";
        // this qry will check app key, if valid then check id
 
         $res = mysqli_fetch_all(mysqli_query($link, $qry), MYSQLI_ASSOC);
-        if($res != null){
-            if(password_verify($data->pass, $res[0]['hasPass'])){
-                $qry = "SELECT onOrOf FROM `wallet` WHERE nsuId = $data->id ";
-                $res1 = mysqli_fetch_all(mysqli_query($link, $qry), MYSQLI_ASSOC);
-                $conObg->detach();
-                $response = array(
-                    "status"=> "ok",
-                    "name" => $res[0]['name'],
-                    "add"  => $res[0]['address'],
-                    "qstn" => $res[0]['qs'],
-                    "flag" => ($res1 == null)? null : $res1[0]['onOrOf']
-                );
-                http_response_code(200);
-                echo json_encode($response);
-            }else{
-                http_response_code(200);
-                echo json_encode(array('status'=> 'invalid'));
-            }
+        if($res != null && password_verify($data->pass, $res[0]['hashPin'])){
+            $conObg->detach();
+            
+            $key = 1234; // set api key here;
+            $xData = new purData;
+            $load = [
+                'key' => $key,
+                'id' => $data->id
+                ];
+                //making post request
+          $bankRes[] = array();
+          $bankRes = $xData->make_req($xData->get_history_url(), $load );
+          
+          http_response_code(200);
+          //echo json_encode($bankRes);
+          print_r($bankRes);
 
             
         }else{  // invalid app key
